Use manual caching instead of proc macro (#12)
* Use manual caching instead of proc macro

required for specifying redis url in tests

+++ src/models/defaults.rs
@@ -26,7 +26,7 @@ pub fn default_session_cleanup_interval_secs() -> u64 {
     600
 }
 
-pub fn default_max_concurrent_sessions() -> u8 {
+pub fn default_max_concurrent_sessions() -> usize {
     5
 }
